requirements.txt
# requirements.txt
python-telegram-bot
python-dotenv
google-genai
openai
asyncio
faster-whisper
ddgs
youtube_transcript_api
python-telegram-bot[job-queue]>=20.0
pytest
pytest-mock
pytest-asyncio
telegramify-markdown
Pillow

